requirements.txt
Flask
python-dotenv
pymongo[srv]
feedparser
python-dateutil
requests
beautifulsoup4
gunicorn